wip: Fix something

+++ lib/domains/forex/models.py
@@ -11,5 +11,5 @@ class ForexDataSchema(pa.DataFrameModel):
     high: float = pa.Field(ge=0, description="Highest price of the forex during the period")
     low: float = pa.Field(ge=0, description="Lowest price of the forex during the period")
     close: float = pa.Field(ge=0, description="Closing price of the forex")
-    volume: float = pa.Field(ge=0, description="Volume of forex traded")
+    volume: int = pa.Field(ge=0, description="Volume of forex traded")
     symbol: str = pa.Field(description="The symbol of the forex. eg. 'JPY=X' for JPY/USD.")
